% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archieve.R
\name{nyt_archieve}
\alias{nyt_archieve}
\title{Get articles from NYT Archieve}
\usage{
nyt_archieve(month = 1, year = 1970, tibble = TRUE)
}
\arguments{
\item{month}{the desired month provided as a wholenumber}

\item{year}{the desired year provided as a wholenumber}

\item{tibble}{if TRUE the API result is formatted and returned as a tibble, else a complete response
is provided}
}
\description{
The function returns a tibble as default, which contains all New York Times articles of the specified
year and month. Articles are available from 1851 and up to the present year and month.
}
\examples{
\dontrun{nyt_archieve(month = 1, year = 1970) #Remember to set token}
}
\seealso{
\link[Rnewyorktimes]{nyt_token}
}
