% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nyt_token}
\alias{nyt_token}
\title{Set API token}
\usage{
nyt_token()
}
\description{
The function stores the token in an environment, which allows the package functions to access the key,
without the need of an token argument, which must be typed each time the function is used. Another reason
to store the token this way is that there is no need to store the token in a variable, which would be
stored in the Rhistory file.
}
\examples{
\dontrun{nyt_token()}
}
