% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnewyorktimes.R
\docType{package}
\name{Rnewyorktimes-package}
\alias{Rnewyorktimes}
\alias{Rnewyorktimes-package}
\title{Rnewyorktimes: Easy access to New York Times articles}
\description{
The package aim at providing a package, which contains all of New York Times APIs. The package contains
a function to retrieve data about:
\itemize{
\item The New York Times Archieve - Contains all articles from 1851 up to today
}
}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

}
